+++ backend/app/api/v1/endpoints/likes.py
@@ -9,6 +9,9 @@ router = APIRouter()
 
 @router.get("/posts/{id}")
 def get_likes_by_post_id():
+    """
+    Retrieve a specified like.
+    """
     try:
         # return crud_likes.CRUD_likes.get_by_comment_id(id)
         return {}
@@ -18,6 +21,9 @@ def get_likes_by_post_id():
 
 @router.get("/comments/{id}")
 def get_likes_by_comment_id():
+    """
+    Retrieve a like by using a specified comment.
+    """
     try:
         # return crud_likes.CRUD_likes.get_by_comment_id(id)
         return {}
@@ -28,6 +34,9 @@ def get_likes_by_comment_id():
 @router.post("/comments")
 @auth_guard("user")
 def manage_like_on_comment(request: Request):
+    """
+    Create a like on a comment.
+    """
     try:
         return crud_likes.CRUD_likes.manage_on_comment()
     except HTTPException:
@@ -37,6 +46,9 @@ def manage_like_on_comment(request: Request):
 @router.post("/posts")
 @auth_guard("user")
 def manage_like_on_post(request: Request):
+    """
+    Create a like on a post.
+    """
     try:
         return crud_likes.CRUD_likes.manage_on_post()
     except HTTPException: