+++ backend/app/mongo/models/users.py
@@ -22,3 +22,16 @@ class User(BaseModel):
     email: EmailStr
     password: str
     isAdmin: bool = False
+
+    class Config:
+        schema_extra = {
+            "example": {
+                "id": 43,
+                "username": "GeneralUserName",
+                "email": "Generic.mail@gmail.com",
+                "password": "password123",
+                "token": "Byxm1iW5J5KMzdbSKwn37CgncQ0MeEhv",
+                "salt": "superSecretKey",
+                "refresh_token": "rMRCZVOfcK"
+            }
+        }