+++ history.md
@@ -125,6 +125,22 @@ Updated CRUD
 * reports
 * users
 
+## 0.2.21 (2022-10-13)
+Updated schemas
+* added schema for reports
+
+## 0.2.22(2022-10-13)
+Updated schemas
+* added schema for comments
+
+## 0.2.23 (2022-10-13)
+Updated models
+* added config for user models
+* improved config for comment models
+
+## 0.2.24 (2022-10-13)
+Updated docstring for api/endpoints
+
 ## 0.3.0 (2022-10-13)
 Implemented the logic for auth signup/login
 * Verified if auth guard was succesfull with real data.